+++ src/components/TickMaps/types.ts
@@ -2,13 +2,11 @@
 export interface TickMapItemProps {
   tickMap: TickMapItemType;
   ID: string;
-  enabled: boolean;
-  setter: any;
-  remover: any;
-  moveUp: any;
-  moveDown: any;
-  createDuplicate: any;
-  context: any;
+  setter: (index: number, value: TickMapItemType) => void;
+  remover: (index: number) => void;
+  moveUp: (index: number) => void;
+  moveDown: (index: number) => void;
+  createDuplicate: (index: number) => void;
 }
 
 export interface TickMapItemType {
@@ -23,4 +21,5 @@ export interface TickMapItemTracker {
   tickMap: TickMapItemType;
   order: number;
   ID: string;
+  isOpen: boolean;
 }